backend/requirements.txt
fastapi>=0.115.2
uvicorn[standard]==0.30.6
python-dotenv>=1.0.1
pydantic>=2.8.2
langchain==0.2.17
langchain-core==0.2.43
langchain-community==0.2.9
langsmith==0.1.147
langchain-openai==0.1.25
langchain-google-genai==1.0.7
google-generativeai>=0.8.3
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
chromadb==0.4.24
PyPDF2==3.0.1
pdfplumber==0.11.4
openai>=1.40.0
python-multipart>=0.0.9

